requirements.txt
asgiref==3.2.10
certifi==2020.6.20
coverage==5.3
Django==3.1
django-debug-toolbar==2.2
django-redis==4.12.1
djangorestframework==3.11.1
mysqlclient==2.0.1
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
